backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
fal-client==0.4.1
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
prometheus-client==0.21.0
sentry-sdk==2.18.0
pytest==8.3.4
pytest-asyncio==0.24.0
